Split IronFish manager to multiple files.
